python/pyproject.toml
[tool.poetry]
name = "Notify Python Example"
version = "0.1.0"
description = ""
authors = ["Stephen Astels <sastels@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
